requirements.txt
appdirs==1.4.4
asttokens==2.4.1
attrs==23.1.0
Brotli==1.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
comet-ml==3.35.3
comm==0.2.0
configobj==5.0.8
debugpy==1.8.0
decorator==5.1.1
defcon==0.10.3
dulwich==0.21.6
everett==3.1.0
executing==2.0.1
filelock==3.9.0
fonttools==4.45.0
freetype-py==2.4.0
fs==2.4.16
fsspec==2023.4.0
idna==3.4
imageio==2.33.0
ipykernel==6.27.0
ipython==8.17.2
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter_client==8.6.0
jupyter_core==5.5.0
lazy_loader==0.3
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.5.8
networkx==3.2.1
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.3
parso==0.8.3
pexpect==4.8.0
Pillow==10.1.0
platformdirs==4.0.0
prompt-toolkit==3.0.41
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.1
python-box==6.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.31.0
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.0
rpds-py==0.13.1
scikit-image==0.22.0
scipy==1.11.4
semantic-version==2.10.0
sentry-sdk==1.36.0
simplejson==3.19.2
six==1.16.0
stack-data==0.6.3
sympy==1.12
tensorfont==0.1.0
tifffile==2023.9.26
torch==2.1.1+cpu
torchaudio==2.1.1+cpu
torchvision==0.16.1+cpu
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
typing_extensions==4.4.0
tzdata==2023.3
ufo-extractor==0.6.0
uharfbuzz==0.37.3
unicodedata2==15.1.0
urllib3==2.1.0
wcwidth==0.2.12
websocket-client==1.3.3
wrapt==1.16.0
wurlitzer==3.0.3
zopfli==0.2.3
